backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
openai==1.3.7
pymongo==4.6.0
djongo==1.3.6
email-validator==2.1.0
python-imap==0.0.11
python-dateutil==2.8.2
requests==2.31.0
PyPDF2==3.0.1
python-multipart==0.0.6
gunicorn==21.2.0
